pyproject.toml
[project]
name = "hmi-ai-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi[standard]>=0.115.12",
    "langchain-core>=0.3.59",
    "langchain-openai>=0.3.16",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "sse-starlette>=2.3.4",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".md",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 80
indent-width = 2

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "E501"]
